Create: 근태 결근

+++ src/main/java/com/goodee/finals/productManage/ProductManageDTO.java
@@ -27,7 +27,7 @@ public class ProductManageDTO {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long pmNum;
 	private Integer pmType;
-	private Integer pmAmount;
+	private Long pmAmount;
 	private LocalDate pmDate = LocalDate.now();
 	
 	@ManyToOne
